feat: started on importer

+++ src/server/virtualmuseum.web/virtualmuseum.web.data/Model/Media/MediaFile.cs
@@ -13,7 +13,7 @@ public class MediaFile
     public string? FileName { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
-    public int DurationInSeconds { get; set; }
+    public float DurationInSeconds { get; set; }
     public MediaType Type { get; set; }
     public string? Url { get; set; }
 }
